% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.binPrb_scoreCI.R
\name{power_via_scoreCIs}
\alias{power_via_scoreCIs}
\title{Power analysis for a test of "H0: binom-prb. CI > delta" via simulation data.}
\usage{
power_via_scoreCIs(
  Ref.Prp = c(rep(0.4, 8), rep(0.2, 8)),
  Test.Prp = c(rep(0.4, 8), rep(0.2, 8)),
  non.inf.bound = 0.15,
  SampleSize.Ref = 176,
  SampleSize.Test = 176,
  SS.prop.Ref = c(rep(0.8 * 1/8, 8), rep(0.2 * 1/8, 8)),
  SS.prop.Test = c(rep(0.8 * 1/8, 8), rep(0.2 * 1/8, 8)),
  adj.prp.min = NULL,
  adj.prp.max = NULL,
  n.adj = 30 * 0.8/0.2,
  conf.level = 0.95,
  itt = 500,
  ci.method = "strMNCI::scoreint.strat.MN",
  str = 16
)
}
\arguments{
\item{Ref.Prp}{A numer. vector of bin. prb. for the reference group.}

\item{Test.Prp}{A numer. vector of bin. prb. for the test group.}

\item{non.inf.bound}{Delta (numeric)}

\item{SampleSize.Ref}{Sample size for the reference group.}

\item{SampleSize.Test}{Sample size for the test group.}

\item{SS.prop.Ref}{A numer. vector of proportions (>0, <1) in the respective strata. If NULL, the sample size will be equally divided for each strata.}

\item{SS.prop.Test}{A numer. vector of proportions (>0, <1) in the respective strata. If NULL, the sample size will be equally divided for each strata.}

\item{adj.prp.min}{NULL}

\item{adj.prp.max}{NULL}

\item{itt}{500}

\item{ci.method}{"strMNCI::scoreint.strat.MN" or "PropCIs::diffscoreci".}

\item{str}{The number of strata.}
}
\description{
Power analysis for a test of "H0: binom-prb. CI > delta" via simulation data.
}
\references{
Eqs.27 and 28 in "Comparative analysis of two rates", MIETTINEN AND NURMINEN, Stat. Med.(1985)
}
